pyproject.toml
[tool.poetry]
name = "equidistant-ml"
version = "0.1.0"
authors = [
    "Ben <btjones-me>"
]
description = "Contains ml for equidistant, an app to estimate journey times"
license = "Not open source"
keywords = [
    "FIXME",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8"
]

homepage = "https://github.com/btjones-me/equidistant-ml"
repository = "https://github.com/btjones-me/equidistant-ml"

readme = "README.md"

[[tool.poetry.source]]
name = "pypi_"
url = "https://pypi.org/simple/"
default = true

#[[tool.poetry.source]]
#name = "pypigetfeed"
#url = "https://pkgs.dev.azure.com/org/_packaging/pypigetfeed/pypi/simple/"
#secondary = true

# Do not manually edit, use poetry add/remove (--dev) to manage dependencies
[tool.poetry.dependencies]
mypy = "^0.782"
python = "^3.8"
pandas = "^1.3.2"
matplotlib = "^3.4.3"
googlemaps = "^4.5.3"
dvc = "^2.6.3"
requests-cache = "^0.7.4"
pyprojroot = "^0.2.0"
tqdm = "^4.62.1"
loguru = "^0.5.3"
openpyxl = "^3.0.7"
python-dotenv = "^0.19.0"

[tool.poetry.dev-dependencies]
assertpy = "^1.0"
bandit = "^1.6.2"
black = "^19.10b0"
coverage = "^5.1"
flake8 = "^3.8.3"
isort = "^5.1.0"
liccheck = "^0.4.7"
pytest-runner = "^5.2"
pytest = "^5.4.3"
tox-gh-actions = "^1.2.0"
tox = "^3.15.2"
toml = "^0.10.1"
# pytest-cov = "^2.10.0"

#[tool.pytest.ini_options]
#addopts = "--cov=equidistant_ml --cov-fail-under=95"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
